fix(tests): fold frontend_pool_allocation config tests into covered file
tests/inductor/test_config.py collided with the upstream PyTorch test
path declared in tests/configs/upstream_tests_beta/inductor-test_config.yaml,
so check_test_coverage.sh treated our tests as already covered by that
unrelated upstream YAML and they never actually ran in CI. Fold the two
tests into test_symbolic_dim_bundle.py, which already has its own
coverage YAML and already exercises frontend_pool_allocation, and
delete the colliding file.
